+++ src/ConversionModel/mustBeConversionModel.ts
@@ -32,51 +32,19 @@
 // POSSIBILITY OF SUCH DAMAGE.
 //
 
-import { roundTo } from "@safelytyped/math-rounding";
-import { rgb } from "culori";
-import type { ConversionModel } from "../ConversionModel/ConversionModel.type";
-import type { CssCmykColorChannelsData } from "./CssCmykColorChannelsData.type";
+import { DEFAULT_DATA_PATH, mustBe, THROW_THE_ERROR, type DataGuaranteeOptions } from "@safelytyped/core-types";
+import type { Color } from "culori";
+import { validateConversionModel } from "./validateConversionModel";
 
-/**
- * convertConversionModelToCmykChannelsData() is a helper method. It converts
- * an instance of the RGB model used by our chosen third-party color
- * conversion package to our preferred data format.
- *
- * based on an algorithm from color-space:
- * https://github.com/colorjs/color-space/blob/master/cmyk.js
- *
- * this algorithm appears to be commonly used by color-conversion websites
- * that predate the color-space package
- *
- * @param input
- * @returns
- */
-export function convertConversionModelToCmykChannelsData(
-    input: ConversionModel
-): CssCmykColorChannelsData
+export function mustBeConversionModel(
+    input: Color,
+    {
+        path = DEFAULT_DATA_PATH,
+        onError = THROW_THE_ERROR
+    }: DataGuaranteeOptions = {}
+)
 {
-    const model = rgb(input);
-
-    const k = Math.min(1 - model.r, 1 - model.g, 1 - model.b);
-    const c = (1 - model.r - k) / (1 - k) || 0;
-    const m = (1 - model.g - k) / (1 - k) || 0;
-    const y = (1 - model.b - k) / (1 - k) || 0;
-
-    return {
-        cyan: round(c * 100),
-        magenta: round(m * 100),
-        yellow: round(y * 100),
-        key: round(k * 100),
-    };
-}
-
-function round(input: number)
-{
-    return Math.abs(
-        roundTo(
-            Math.round,
-            0,
-            input,
-        )
-    );
+    return mustBe(input, { onError })
+        .next((x) => validateConversionModel(x, { path}))
+        .value();
 }