+++ src/ConversionModel/validateConversionModel.ts
@@ -32,51 +32,27 @@
 // POSSIBILITY OF SUCH DAMAGE.
 //
 
-import { roundTo } from "@safelytyped/math-rounding";
-import { rgb } from "culori";
-import type { ConversionModel } from "../ConversionModel/ConversionModel.type";
-import type { CssCmykColorChannelsData } from "./CssCmykColorChannelsData.type";
+import { DEFAULT_DATA_PATH, UnsupportedTypeError, type AppErrorOr, type DataValidatorOptions } from "@safelytyped/core-types";
+import type { Color } from "culori";
+import type { ConversionModel } from "./ConversionModel.type";
+import { SUPPORTED_CONVERSION_MODEL_MODES } from "./SUPPORTED_CONVERSION_MODEL_MODES";
 
-/**
- * convertConversionModelToCmykChannelsData() is a helper method. It converts
- * an instance of the RGB model used by our chosen third-party color
- * conversion package to our preferred data format.
- *
- * based on an algorithm from color-space:
- * https://github.com/colorjs/color-space/blob/master/cmyk.js
- *
- * this algorithm appears to be commonly used by color-conversion websites
- * that predate the color-space package
- *
- * @param input
- * @returns
- */
-export function convertConversionModelToCmykChannelsData(
-    input: ConversionModel
-): CssCmykColorChannelsData
+export function validateConversionModel(
+    input: Color,
+    {
+        path = DEFAULT_DATA_PATH
+    }: DataValidatorOptions = {}
+): AppErrorOr<ConversionModel>
 {
-    const model = rgb(input);
+    if (SUPPORTED_CONVERSION_MODEL_MODES.includes(input.mode)) {
+        return input as ConversionModel;
+    }
 
-    const k = Math.min(1 - model.r, 1 - model.g, 1 - model.b);
-    const c = (1 - model.r - k) / (1 - k) || 0;
-    const m = (1 - model.g - k) / (1 - k) || 0;
-    const y = (1 - model.b - k) / (1 - k) || 0;
-
-    return {
-        cyan: round(c * 100),
-        magenta: round(m * 100),
-        yellow: round(y * 100),
-        key: round(k * 100),
-    };
-}
-
-function round(input: number)
-{
-    return Math.abs(
-        roundTo(
-            Math.round,
-            0,
-            input,
-        )
-    );
+    return new UnsupportedTypeError({
+        public: {
+            dataPath: path,
+            expected: "supported ConversionModel object",
+            actual: input.mode,
+        }
+    });
 }